--- src/main/java/phanes/replay/gathering/domain/Role.java
@@ -1,5 +1,0 @@
-package phanes.replay.gathering.domain;
-
-public enum Role {
-    HOST, MEMBER
-}
