+++ src/main/java/phanes/replay/gathering/domain/enums/Role.java
@@ -1,0 +1,5 @@
+package phanes.replay.gathering.domain.enums;
+
+public enum Role {
+    HOST, MEMBER
+}
