+++ src/main/java/phanes/replay/user/service/UserService.java
@@ -6,7 +6,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import phanes.replay.gathering.domain.*;
+import phanes.replay.gathering.domain.GatheringComment;
+import phanes.replay.gathering.domain.GatheringMember;
+import phanes.replay.gathering.domain.GatheringScheduleView;
+import phanes.replay.gathering.domain.ParticipatingGatheringView;
+import phanes.replay.gathering.domain.enums.Role;
 import phanes.replay.gathering.service.GatheringCommentService;
 import phanes.replay.gathering.service.GatheringMemberService;
 import phanes.replay.gathering.service.GatheringService;
